refactor: Move types around
Almost got rid of repn-invoker... soon
