+++ src/repm-invoker.ts
@@ -1,65 +1,6 @@
-import type {Puller} from './puller';
-import type {Pusher} from './pusher';
 import type * as dag from './dag/mod';
-import type * as db from './db/mod';
-import type {JSONValue} from './json';
 
 export type OpenResponse = {clientID: string; store: dag.Store};
-
-type TransactionRequest = {
-  transactionId: number;
-};
-
-export type ScanRequest = TransactionRequest & {
-  opts?: db.ScanOptions;
-  receiver: (
-    primaryKey: string,
-    secondaryKey: string | null,
-    value: Uint8Array,
-  ) => void;
-};
-export type ScanResponse = unknown;
-
-export type RebaseOpts = {
-  basis: string;
-  original: string;
-};
-
-export type OpenTransactionRequest = {
-  name?: string;
-  args?: JSONValue;
-  rebaseOpts?: RebaseOpts;
-};
-
-export type CloseTransactionResponse = unknown;
-
-// The changed keys in different indexes. The key of the map is the index name.
-// "" is used for the primary index.
-export type ChangedKeysMap = Map<string, string[]>;
-
-export type CommitTransactionResponse = {
-  ref: string;
-  changedKeys: ChangedKeysMap;
-};
-
-export type BeginPullRequest = {
-  pullURL: string;
-  pullAuth: string;
-  schemaVersion: string;
-  puller: Puller;
-};
-
-export type BeginPullResponse = {
-  httpRequestInfo: HTTPRequestInfo;
-  syncHead: string;
-};
-
-export type TryPushRequest = {
-  pushURL: string;
-  pushAuth: string;
-  schemaVersion: string;
-  pusher: Pusher;
-};
 
 export function assertHTTPRequestInfo(
   // eslint-disable-next-line
@@ -78,26 +19,4 @@ export function assertHTTPRequestInfo(
 export type HTTPRequestInfo = {
   httpStatusCode: number;
   errorMessage: string;
-};
-
-export type MaybeEndPullRequest = {
-  requestID: string;
-  syncHead: string;
-};
-
-/**
- * ReplayMutation is used int the RPC between EndPull so that we can replay
- * mutations ontop of the current state. It is never exposed to the public.
- */
-export type ReplayMutation = {
-  id: number;
-  name: string;
-  args: JSONValue;
-  original: string;
-};
-
-export type MaybeEndPullResponse = {
-  replayMutations?: ReplayMutation[];
-  syncHead: string;
-  changedKeys: ChangedKeysMap;
 };
