+++ src/sync/mod.ts
@@ -3,3 +3,6 @@ export {maybeEndPull, beginPull} from './pull';
 export {push} from './push';
 export {newRequestID} from './request-id';
 export {SYNC_HEAD_NAME} from './sync-head-name';
+export {validateRebase} from './validate-rebase';
+export type {RebaseOpts} from './validate-rebase';
+export type {ChangedKeysMap} from './pull';
